% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomial_helper}
\alias{binomial_helper}
\title{TODO}
\usage{
binomial_helper(raw, levels, type = c("prob", "class"),
  threshold = 0.5)
}
\arguments{
\item{raw}{A numeric vector of probabilities.}

\item{levels}{A two-element character vector of class names. The first
element is the name of the baseline class, and the second element is
the name of the positive event. Probabilities specified in \code{raw} are
assumed to be for the positive event.}

\item{type}{TODO}

\item{threshold}{TODO}
}
\description{
TODO
}
