% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-glmnet-multi.R
\name{multi_predict.glmnet}
\alias{multi_predict.glmnet}
\title{Safe predictions for cross-validated glmnet objects}
\usage{
\method{multi_predict}{glmnet}(object, new_data, type = c("response",
  "class", "prob"), ..., params = NULL)
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{type}{TODO}

\item{...}{Unused. \code{safe_predict()} checks that all arguments in \code{...} are
evaluated via the \code{ellipsis} package. The idea is to prevent silent errors
when arguments are mispelled. This feature is experimental and feedback
is welcome.}

\item{params}{\itemize{
\item works with a single value for the penalty. otherwise CV is gonna be
inconsistent a wild pain in the rear.
\itemize{
\item works with a vector of values for the penalty
\item in the future will work with a tibble defining a parameter grid
}
}}
}
\description{
Safe predictions for cross-validated glmnet objects
}
