% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safepredict-package.R
\name{multi_predict}
\alias{multi_predict}
\alias{multi_predict.default}
\title{Safely predict from many fits at once}
\usage{
multi_predict(object, ...)

\method{multi_predict}{default}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Experimental and untested, not recommended for general use.
}
