% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{positive_class}
\alias{positive_class}
\title{Determine the positive class of a logistic regression fit with \code{glm}}
\usage{
positive_class(object)
}
\arguments{
\item{object}{TODO}
}
\value{
TODO
}
\description{
Determine the positive class of a logistic regression fit with \code{glm}
}
