% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.0.R
\name{safe_predict.C5.0}
\alias{safe_predict.C5.0}
\title{Safe predictions from a C5 classification tree}
\usage{
\method{safe_predict}{C5.0}(object, new_data, type = c("prob", "class"),
  ..., threshold = 0.5)
}
\arguments{
\item{object}{A \code{C5.0} object produced by a call to \code{\link[C50:C5.0]{C50::C5.0()}}.}

\item{new_data}{TODO}

\item{type}{TODO}

\item{...}{Unused. \code{safe_predict()} checks that all arguments in \code{...} are
evaluated via the \code{ellipsis} package. The idea is to prevent silent errors
when arguments are mispelled. This feature is experimental and feedback
is welcome.}

\item{threshold}{A number between \code{0} and \code{1} to use as a threshold for
classification. When the class probability for the class corresponding
to a positive event is greater than the threshold, the event will be
classified as positive. Defaults to \code{0.5}.}
}
\description{
Safe predictions from a C5 classification tree
}
