% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-cv-glmnet.R
\name{safe_predict.cv.glmnet}
\alias{safe_predict.cv.glmnet}
\title{Safe predictions for cross-validated glmnet objects}
\usage{
\method{safe_predict}{cv.glmnet}(object, new_data, type = c("response",
  "class", "prob", "link"), ..., rule = c("1-se", "min"),
  threshold = 0.5)
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{type}{TODO}

\item{...}{TODO}

\item{rule}{TODO}

\item{threshold}{TODO}
}
\description{
Safe predictions for cross-validated glmnet objects
}
