% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-glm.R
\name{safe_predict.glm}
\alias{safe_predict.glm}
\title{Safe predictions from a generalized linear model}
\usage{
\method{safe_predict}{glm}(object, new_data, type = c("link", "conf_int",
  "response", "class", "prob"), ..., std_error = FALSE, level = 0.95,
  threshold = 0.5)
}
\arguments{
\item{object}{A \code{glm} object returned from a call to \code{\link[stats:glm]{stats::glm()}}.}

\item{new_data}{TODO}

\item{type}{What kind of predictions to return. Which predictions are
available depends on the family of \code{object}.

\code{"link"} and \code{"conf_int"} are available for all families. \code{"link"}
produces numeric predictions on the linear predictor scale. \code{"conf_int"}
produces numeric predictions on the response scale and corresponding
confidence bounds.
\itemize{
\item \code{"response"} results in a numeric prediction on the response scale
and is available for families:
\itemize{
\item \code{gaussian}
\item \code{Gamma}
\item \code{inverse.gaussian}
\item \code{poisson}
\item \code{quasipoisson}
\item \code{quasi}
}
\item \code{"class"} results in hard class predictions and is only available for
\code{binomial} and \code{quasibinomial} families
\item \code{"prob"} results in class predictions for each class and is only
available for \code{binomial} and \code{quasibinomial} families
}

Default is \code{"link"}.}

\item{...}{Unused. \code{safe_predict()} checks that all arguments in \code{...} are
evaluated via the \code{ellipsis} package. The idea is to prevent silent errors
when arguments are mispelled. This feature is experimental and feedback
is welcome.}

\item{std_error}{Logical indicating whether or not calculate standard
errors for the fit at each point. Not available for all models, and can
be computationally expensive to compute. The standard error is \emph{always}
the standard error for the mean, and \emph{never} the standard error for
predictions. Standard errors are returned in a column called \code{.std_error}.
Defaults to \code{FALSE}.}

\item{level}{A number strictly between \code{0} and \code{1} to use as the
confidence level when calculating confidence and prediction intervals.
Setting \code{level = 0.90} correspondings to a 90 percent confidence interval.
Ignored except when \code{type = "conf_int"} or \code{type = "pred_int"}. Defaults
to \code{0.95}.}

\item{threshold}{A number between \code{0} and \code{1} to use as a threshold for
classification. When the class probability for the class corresponding
to a positive event is greater than the threshold, the event will be
classified as positive. Defaults to \code{0.5}.}
}
\description{
Safe predictions from a generalized linear model
}
\details{
For GLMs, standard errors can only be calculated when
\code{type = "link"}.
}
