% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{safe_predict.lm}
\alias{safe_predict.lm}
\title{Safe predictions from a linear model}
\usage{
\method{safe_predict}{lm}(object, newdata, type = c("response",
  "conf_int", "pred_int"), se_fit = TRUE, level = 0.95, ...)
}
\arguments{
\item{object}{An `lm` object returned from a call to [stats::lm()].}

\item{newdata}{A dataset as a [data.frame] object. In theory this should
work with any object accepted by the [stats::predict.lm()] `newdata`
argument that can be reasonably coerced into a tibble.}

\item{type}{What kind of predictions to return. Options are:
- `"response"` (default): Standard predictions from linear regression.
- `"conf_int"`: Fitted values plus a confidence interval for the fit.
- `"pred_int"`: Predictions with accompanying prediction interval.}

\item{se_fit}{Logical indicating whether or not to also calculate standard
errors for the fit at each point. These standard errors do not including
the residual variance. That is, they can be used to calculate confidence
intervals but not prediction intervals.}

\item{level}{A number between 0 and 1 to use as the confidence level to
use when calculating confidence and prediction intervals.}

\item{...}{Unused. TODO: boilerplate for this.}
}
\value{
A [tibble::tibble()] with one row for each row of `newdata`.
  predictions for observations with missing data will be `NA`. Predictions
  are contained in the
  boilerplate about safe_predict_guarantees
}
\description{
Safe predictions from a linear model
}
\details{
*What is the difference between confidence intervals and prediction
  intervals*? TODO.

Do not use on model objects that only subclass `lm`. Will error.
}
