% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-mlm.R
\name{safe_predict.mlm}
\alias{safe_predict.mlm}
\title{Safe predictions from a multiple linear model object}
\usage{
\method{safe_predict}{mlm}(object, new_data, type = c("response"), ...)
}
\arguments{
\item{object}{An \code{mlm} object returned from a call to \code{\link[stats:lm]{stats::lm()}}.}

\item{new_data}{\strong{Required}. A data frame or matrix containing the
necessary predictors.}

\item{type}{What kind of predictions to return. Options are:
\itemize{
\item \code{"response"} (default): Standard predictions from multiple regression.
}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each row of \code{new_data}.
Predictions for observations with missing data will be \code{NA}. Returned
tibble has different columns depending on \code{type}:
\itemize{
\item \code{"response"}:
\itemize{
\item univariate outcome: \code{.pred} (numeric)
\item multivariate outcomes: \code{.pred_{outcome name}} (numeric) for each
outcome
}
\item \code{"class"}: \code{.pred_class} (factor)
\item \code{"prob"}: \code{.pred_{level}} columns (numerics between 0 and 1)
\item \code{"link"}: \code{.pred} (numeric)
\item \code{"conf_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
\item \code{"pred_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
}

If you request standard errors with \code{std_error = TRUE}, an additional
column \code{.std_error}.

For interval predictions, the tibble has additional attributes \code{level}
and \code{interval}. The \code{level} is the same as the \code{level} argument and is
between 0 and 1. \code{interval} is either \code{"confidence"} or \code{"prediction"}.
Some models may also set a \code{method} attribute to detail the method
used to calculate the intervals.
}
\description{
Safe predictions from a multiple linear model object
}
\section{Estimating uncertainty}{


\code{\link[stats:predict.mlm]{stats::predict.mlm()}} provides neither confidence nor prediction
intervals, although there is not theoretical issue with calculating
these.

At some point in the future we may implement these intervals within
\code{safepredict}. If you are interested in this, you can move intervals
for \code{mlm} objects up the priority list by opening an issue on
\href{https://github.com/alexpghayes/safepredict}{Github}.
}

\examples{

fit <- lm(cbind(hp, mpg) ~ ., mtcars)

safe_predict(fit, mtcars)

mt2 <- mtcars
diag(mt2) <- NA

safe_predict(fit, mt2)

}
