% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-model_fit.R
\name{safe_predict.model_fit}
\alias{safe_predict.model_fit}
\title{Safely predict from a model_fit object}
\usage{
\method{safe_predict}{model_fit}(object, new_data, type = NULL, ...,
  opts = list())
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{...}{Unused.}
}
\value{
TODO
}
\description{
WHY THIS: need a special case where \code{type = "raw"} is acceptable
}
