% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-prcomp.R
\name{safe_predict.prcomp}
\alias{safe_predict.prcomp}
\alias{safe_predict.princomp}
\title{Safe predictions from a prcomp object}
\usage{
\method{safe_predict}{prcomp}(object, new_data, type = "response", ...)

\method{safe_predict}{princomp}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{A \code{prcomp} object returned from a call to \code{\link[stats:prcomp]{stats::prcomp()}}
or a \code{princomp} object returned from a call to \code{\link[stats:princomp]{stats::princomp()}}.}

\item{new_data}{TODO: what is acceptable here?}

\item{type}{What kind of predictions to return. There is only one option:
\itemize{
\item \code{"response"} (default): Standard PCA rotation without scaling
}}
}
\value{
What columns you get back, and the details of the computation
\itemize{
\item rotation details
\item PC1, ..., PCn
}
}
\description{
Safe predictions from a prcomp object
}
\concept{unsupervised}
