% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safepredict-package.R
\name{safe_predict}
\alias{safe_predict}
\title{Safely predict from a model object}
\usage{
safe_predict(object, new_data, type = NULL, opts = list(),
  level = 0.95, std_error = FALSE, ...)
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{...}{Unused.}
}
\value{
TODO
}
\description{
Safely predict from a model object
}
