% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safepredict-package.R
\docType{package}
\name{safepredict-package}
\alias{safepredict}
\alias{safepredict-package}
\title{safepredict: Consistent prediction following tidymodels principles}
\description{
Provides the safe_predict() generic, a consistent
    interface to prediction that meets that tidymodels prediction
    specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/safepredict}
  \item Report bugs at \url{https://github.com/alexpghayes/safepredict/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
