% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-time-series.R
\name{suggest_sweep}
\alias{suggest_sweep}
\alias{safe_predict.ar}
\alias{safe_predict.Arima}
\alias{safe_predict.Arima0}
\alias{safe_predict.HoltWinters}
\alias{safe_predict.StructTS}
\title{Predictions from time series models}
\usage{
suggest_sweep(object, new_data, type, ...)

\method{safe_predict}{ar}(object, new_data, type, ...)

\method{safe_predict}{Arima}(object, new_data, type, ...)

\method{safe_predict}{Arima0}(object, new_data, type, ...)

\method{safe_predict}{HoltWinters}(object, new_data, type, ...)

\method{safe_predict}{StructTS}(object, new_data, type, ...)
}
\arguments{
\item{object}{Ignored.}

\item{new_data}{Ignored.}

\item{type}{Ignored.}

\item{...}{Ignored.}
}
\description{
\code{safepredict} does not support time series models. We recommend using
\link[sweep:sweep_package]{sweep}.
}
