% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-cv-glmnet-multi.R
\name{multi_predict.cv.glmnet}
\alias{multi_predict.cv.glmnet}
\title{Safe predictions for cross-validated glmnet objects}
\usage{
\method{multi_predict}{cv.glmnet}(object, new_data, type = c("response",
  "class", "prob", "link"), ..., params = NULL)
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{type}{TODO}

\item{...}{TODO}

\item{params}{TODO}
}
\value{
TODO
}
\description{
Safe predictions for cross-validated glmnet objects
}
\concept{glmnet}
