% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-glm.R
\name{safe_predict.glm}
\alias{safe_predict.glm}
\title{Safe predictions from a generalized linear model}
\usage{
\method{safe_predict}{glm}(object, new_data, type = c("link", "conf_int",
  "response", "class", "prob"), std_error = FALSE, level = 0.95,
  threshold = 0.5, ...)
}
\arguments{
\item{object}{A \code{glm} object returned from a call to \code{\link[stats:glm]{stats::glm()}}.}

\item{new_data}{A dataset as a \link{data.frame} or similar object. Must be
coercable to a tibble using \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}

\item{type}{What kind of predictions to return. Which predictions are
available depends on the family of \code{object}.

\code{"link"} and \code{"conf_int"} are available for all families. \code{"link"}
produces numeric predictions on the linear predictor scale. \code{"conf_int"}
produces numeric predictions on the response scale and corresponding
confidence bounds.
\itemize{
\item \code{"response"} results in a numeric prediction on the response scale
and is available for families:
\itemize{
\item \code{gaussian}
\item \code{Gamma}
\item \code{inverse.gaussian}
\item \code{poisson}
\item \code{quasipoisson}
\item \code{quasi}
}
\item \code{"class"} results in hard class predictions and is only available for
\code{binomial} and \code{quasibinomial} families
\item \code{"prob"} results in class predictions for each class and is only
available for \code{binomial} and \code{quasibinomial} families
}

Default is \code{"link"}.}

\item{level}{A number between \code{0} and \code{1} to use as the confidence level
when calculating confidence and prediction intervals. Ignored
otherwise.}

\item{threshold}{A number between \code{0} and \code{1} to use as a threshold for
classification. When the class probability for the class corresponding
to a positive event is greater than the threshold, the event will be
classified as positive. Defaults to \code{0.5}. See \code{\link[=positive_class]{positive_class()}} for
assistance determining which class is considered the positive class.}

\item{...}{Unused.}

\item{se_fit}{Logical indicating whether or not to also calculate standard
errors for the fit at each point. These standard errors do not include
the residual variance. Ignored when calculating a confidence or prediction
interval.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each row of \code{newdata}.
Predictions for observations with missing data will be \code{NA}. Returned
tibble has different columns depending on \code{type}:
\itemize{
\item \code{"response"}:
\itemize{
\item univariate outcome: \code{.pred} (numeric)
\item multivariate outcomes: \code{.pred_{outcome name}} (numeric) for each
outcome
}
\item \code{"class"}: \code{.pred_class} (factor)
\item \code{"prob"}: \code{.pred_{level}} columns (numerics between 0 and 1)
\item \code{"link"}: \code{.pred} (numeric)
\item \code{"conf_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
\item \code{"pred_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
}

For interval predictions, the tibble has additional attributes \code{level}
and \code{interval}. The \code{level} is the same as the \code{level} argument and is
between 0 and 1. \code{interval} is either \code{"confidence"} or \code{"prediction"}.
Some models may also set a \code{method} attribute to detail the method
used to calculate the intervals.
}
\description{
Safe predictions from a generalized linear model
}
\details{
For GLMs, standard errors can only be calculated when
\code{type = "link"}.
}
