% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-glmnet.R
\name{safe_predict.glmnet}
\alias{safe_predict.glmnet}
\title{Safe predictions for glmnet objects}
\usage{
\method{safe_predict}{glmnet}(object, new_data, type = c("response",
  "class", "prob", "link"), penalty = NULL, threshold = 0.5, ...)
}
\arguments{
\item{object}{TODO}

\item{new_data}{TODO}

\item{type}{TODO}

\item{penalty}{Unlike \code{\link[=safe_predict.cv.glmnet]{safe_predict.cv.glmnet()}}, here you can explicitly
set a value of \code{penalty}. In the CV version you have to pick one of the
cross-validated versions of the penalty.}

\item{...}{Unused.}

\item{new_data}{A dataset as a \link{data.frame} or similar object. Must be
coercable to a tibble using \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}

\item{se_fit}{Logical indicating whether or not to also calculate standard
errors for the fit at each point. These standard errors do not include
the residual variance. Ignored when calculating a confidence or prediction
interval.}

\item{level}{A number between \code{0} and \code{1} to use as the confidence level
when calculating confidence and prediction intervals. Ignored
otherwise.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each row of \code{newdata}.
Predictions for observations with missing data will be \code{NA}. Returned
tibble has different columns depending on \code{type}:
\itemize{
\item \code{"response"}:
\itemize{
\item univariate outcome: \code{.pred} (numeric)
\item multivariate outcomes: \code{.pred_{outcome name}} (numeric) for each
outcome
}
\item \code{"class"}: \code{.pred_class} (factor)
\item \code{"prob"}: \code{.pred_{level}} columns (numerics between 0 and 1)
\item \code{"link"}: \code{.pred} (numeric)
\item \code{"conf_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
\item \code{"pred_int"}: \code{.pred}, \code{.pred_lower}, \code{.pred_upper} (all numeric)
}

For interval predictions, the tibble has additional attributes \code{level}
and \code{interval}. The \code{level} is the same as the \code{level} argument and is
between 0 and 1. \code{interval} is either \code{"confidence"} or \code{"prediction"}.
Some models may also set a \code{method} attribute to detail the method
used to calculate the intervals.
}
\description{
Safe predictions for glmnet objects
}
