% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-nls.R
\name{safe_predict.nls}
\alias{safe_predict.nls}
\title{Safe predictions from an nls object}
\usage{
\method{safe_predict}{nls}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{An \code{nls} object returned from a call to \code{\link[stats:nls]{stats::nls()}}.}

\item{new_data}{\strong{Required}. A data frame containing predictors.}

\item{type}{What kind of predictions to return. Options are:
\itemize{
\item \code{"response"} (default): Standard predictions from non-linear regression.
}}
}
\value{
TODO
}
\description{
Safe predictions from an nls object
}
\section{Determining uncertainty in predictions}{


Note that \code{\link[stats:predict.nls]{stats::predict.nls()}} has an \code{se.fit} argument, but it is
currently ignored. There is no build-in capability in base R to determine
the uncertainty in the predictions from \code{nls}.

In practice, there are two options to get these:
\enumerate{
\item Bootstrapping (recommended)
\item The Delta Method
}

\strong{Bootstrapping}: We recommend using the \code{rsample} package for
bootstrapping, in particular \code{\link[rsample:bootstraps]{rsample::bootstraps()}}. See\preformatted{vignette("bootstrapping", package = "safepredict")
}

for worked examples.

\strong{Delta Method}: Some people seem to be happy to use the delta method,
and others claim it is numerically unstable. Two options include
\code{\link[car:DeltaMethod]{car::DeltaMethod()}} and \code{\link[emdbook:deltavar]{emdbook::deltavar()}}.
}

\examples{

fit <- nls(demand ~ SSasympOrig(Time, A, lrc), data = BOD)
safe_predict(fit, BOD)

fit2 <- nls(mpg ~ k / wt + b, mtcars, start = list(k = 1, b = 0))
safe_predict(fit2, mtcars)

}
