% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safepredict-package.R
\docType{package}
\name{safepredict-package}
\alias{safepredict}
\alias{safepredict-package}
\title{safepredict: Consistent prediction following tidymodels principles}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/safepredict}
  \item Report bugs at \url{https://github.com/alexpghayes/safepredict/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
