% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-time-series.R
\name{suggest_sweep}
\alias{suggest_sweep}
\alias{safe_predict.ar}
\alias{safe_predict.Arima}
\alias{safe_predict.Arima0}
\alias{safe_predict.HoltWinters}
\alias{safe_predict.StructTS}
\title{Predictions from time series models}
\usage{
suggest_sweep(object, ...)

\method{safe_predict}{ar}(object, ...)

\method{safe_predict}{Arima}(object, ...)

\method{safe_predict}{Arima0}(object, ...)

\method{safe_predict}{HoltWinters}(object, ...)

\method{safe_predict}{StructTS}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Gracefully error out and suggest using \code{sweep} instead, with some code
examples of what that looks like.
}
